pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentops"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langserve[all]>=0.3,<0.4",
    "langchain-anthropic",
    "langchain-chroma",
    "langchain-text-splitters",
    "langchain-tavily",
    "langsmith>=0.1",
    "arq>=0.26",
    "redis[hiredis]>=5.0",
    "httpx>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "psycopg[binary,pool]>=3.1",
    "chromadb>=0.5",
    "chardet>=5.2",
    "cryptography>=42",
    "pyjwt>=2.8",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-prometheus>=0.46",
    "prometheus-client>=0.20",
    "githubkit>=0.14.7",
    "strawberry-graphql[fastapi]>=0.243",
]

[project.scripts]
agentops-codegen = "agentops.codegen:main"

[dependency-groups]
dev = ["ruff>=0.6", "ty>=0.0.1a1", "pre-commit>=3.8"]
test = [
    "codebase_search",
    "critic",
    "investigator",
    "web_search",
    "writer",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-xdist>=3.6",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.1",
    "fakeredis[aioredis]>=2.26",
    "testcontainers[redis,postgres]>=4.8",
    "asgi-lifespan>=2.1",
    "langchain-tests>=0.3",
    "playwright>=1.44",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "ANN", "N", "B", "S", "T20"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "ANN"]
"scripts/**" = ["T20"]
"src/agentops/graphql/schema.py" = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentops"]

[tool.uv.sources]
codebase_search = {path = "agents/codebase_search", editable = true}
critic = {path = "agents/critic", editable = true}
investigator = {path = "agents/investigator", editable = true}
web_search = {path = "agents/web_search", editable = true}
writer = {path = "agents/writer", editable = true}
